third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/css-cascade-upgrade-from-another-tree-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: host.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL css-cascade-upgrade-from-another-tree Uncaught TypeError: host.createShadowRoot is not a function
Harness: the test ran to completion.
